maintain existing child id whenever possible (#2101)
Child IDs are assigned when receiving a Child ID Request.  Prior to this
commit, a new Child ID is assigned whenever the child state is not valid.
A router stores state about its attached children in non-volatile memory.
On a reboot, the router attempts to restore its child links.  As a result,
it is possible for the child to be assigned a new Child ID while it is
being restored.

With this commit, a router only assigns a new Child ID if there is no
existing Child ID state for the given child.