enable retransmission of network data to SEDs (#2107)
The Thread Specification requires the router to repeatedly transmit
new Thread Network Data to the SED until it receive an explicit
acknolwedgment from the SED.  Prior to this commit, when network data
changes, a router attempts to update the SED by sending a single MLE
Data Response message.  If the message is lost for any reason (e.g. no
available message buffers, link-layer delivery failure, etc.), the
router will not attempt to send the MLE Data Response again to update
the SED's network data.

This commit makes the following changes:

- The router communicates updated network data in an MLE Child Update
  Request message, requiring the SED to send an MLE Child Update
  Response message with the Leader Data TLV.

- The router repeatedly attempts to transmit new network data to the
  SED until it receives a Leader Data TLV with the updated version.