[link-quality] track MAC frame and message error rates (#2520)
This commit implement a new feature to allow the tracking of tx error
rates at two layers:

- At PHY/MAC layer for MAC frame transmissions (frame tx error rate):
  Error rate is maintained per link/neighbor. A frame transmission
  attempt is considered failed if the frame is sent over the channel,
  but no acknowledgment frame is received from the neighbor.

- Above MAC layer for message transmissions (message error rate):
  A larger (IPv6) message may be fragmented and sent as multiple MAC
  frames. The message transmission is considered failed, if any of
  its fragments fail after all retry attempts.  Message error rate is
  maintained per link/neighbor and includes both direct and indirect
  message transmissions.

Configuration option `OPENTHREAD_CONFIG_ENABLE_TX_ERROR_RATE_TRACKING`
can be used to enable/disable this feature. By default it is enabled.

`otChildInfo` and `otNeighborInfo` structs are updated to provide the
error rates for a child or neighbor.