[MARBLE-842] create a copy of Nordic third party driver
This commit creates a copy of Nodrdic third party driver files  under
`NordicSemiconductorV1` folder which is then used for  Nest product
builds.

There are multiple changes on OpenThread GitHub related Nordic driver.
However, Nest devices are currently running on earlier version of
driver.  MARBLE-805 is the JIRA to update/sync the Nordic driver for
Nest and use in  nlOpenThread which blocks the OpenThread GitHub sync.

This commit aims to unblock the sync and ensure git history of changes
are maintained. This should be reverted once MARBLE-805 is done.