[settings] add posix APIs to securely store sensitive settings
All the settings are written into a local file on the posix platform.
It is easy for others to get the sensitive infomation, such as master key,
from the local file. This PR adds new posix APIs `otPlatSecureSettingsXxx`
to posix platform to allow the vendor to store the sensitive settings to a
secure area.